Fixed issues raised on training day 1

+++ application/controllers/home.php
@@ -97,6 +97,7 @@ class Home extends MY_Controller
     //format the graph here
     $facility_code=$this -> session -> userdata('facility_id'); 
     $facility_stock_=facility_stocks::get_facility_stock_amc($facility_code);
+
 	$facility_stock_count=count($facility_stock_);
     $graph_data=array();
 	$graph_data=array_merge($graph_data,array("graph_id"=>'container'));
@@ -109,7 +110,7 @@ class Home extends MY_Controller
 	foreach($facility_stock_ as $facility_stock_):
 		$graph_data['graph_categories']=array_merge($graph_data['graph_categories'],array($facility_stock_['commodity_name']));	
 		$graph_data['series_data']['Current Balance']=array_merge($graph_data['series_data']['Current Balance'],array((float) $facility_stock_['pack_balance']));
-        $graph_data['series_data']['AMC']=array_merge($graph_data['series_data']['AMC'],array((int) $facility_stock_['amc']));	
+        $graph_data['series_data']['AMC']=array_merge($graph_data['series_data']['AMC'],array((float) $facility_stock_['amc']));	
 
 	endforeach;
 	//create the graph here
